requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
fastapi==0.115.12
